создание фротенд и настройка работы api для tasks

+++ backend/app/routers/TaskRouter.py
@@ -25,10 +25,16 @@ def create_task(payload: TaskCreate, taskService: TaskService = Depends()):
 
     return new_task
 
+# добавить endpoint /tasks/{id}/start и POST /tasks/{id}/stop)
+
 @TaskRouter.delete("/{task_id}")
 def delete(task_id: int, taskService: TaskService = Depends()):
     task = taskService.get_task_by_id(task_id)
     check_is_none(task, task_id)
+    
+    if task.on_control == 'on':
+        raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE,detail=f"Can not be deleted while ON")
+   
     taskService.delete(task)
 
     #to do добавить логи childs что parent был удален
@@ -37,20 +43,43 @@ def delete(task_id: int, taskService: TaskService = Depends()):
 @TaskRouter.get("/{task_id}")
 def get_task_by_id(task_id: int, taskService: TaskService = Depends()):
     task = taskService.get_task_by_id(task_id)
-    check_is_none(task, task_id)
-    
+    check_is_none(task, task_id)  
     return task
 
 @TaskRouter.get("")
 def get_tasks(taskService: TaskService = Depends()):
     return taskService.get_all()
 
+@TaskRouter.put("/{task_id}/on")
+def get_tasks(task_id: int, taskService: TaskService = Depends()):
+    task = taskService.get_task_by_id(task_id)
+    check_is_none(task, task_id)
+
+    if task.on_control == 'off':
+        task.on_control = 'on'
+    else: 
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT,detail=f"You try to ON while it is already ON")
+    
+    taskService.update(task)
+
+@TaskRouter.put("/{task_id}/off")
+def get_tasks(task_id: int, taskService: TaskService = Depends()):
+    task = taskService.get_task_by_id(task_id)
+    check_is_none(task, task_id)
+
+    if task.on_control == 'on':
+        task.on_control = 'off'
+    else: 
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT,detail=f"You try to OFF while it is already OFF")
+    
+    taskService.update(task)
+
 @TaskRouter.put("/{task_id}")
 def update_task(task_id: int, data: dict, taskService: TaskService = Depends()):
     task = taskService.get_task_by_id(task_id)
     check_is_none(task, task_id)
 
-    old_control = task.control
+    old_control = task.on_control
     new_control = data.get('control')
 
     if new_control == 'off' and old_control == 'on':
@@ -74,122 +103,3 @@ def check_is_none(object, param):
     if object.status == TaskStatusEnum.RUNNING:
         logger.warning(f"Task id '{object.task_id}' is running can not be updated")
         raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE,detail=f"Task id '{object.task_id}' is running can not be updated")
-
-
-
-
-
-# @app.get("/tasks", response_model=list[Task])
-# def list_tasks() -> list[Task]:
-#     with get_connection() as connection:
-#         rows = connection.execute(text("SELECT * FROM tasks ORDER BY id ASC")).mappings().all()
-#     return [_task_from_row(row) for row in rows]
-
-
-# @app.get("/tasks/{task_id}", response_model=Task)
-# def get_task(task_id: int) -> Task:
-#     with get_connection() as connection:
-#         row = connection.execute(
-#             text("SELECT * FROM tasks WHERE id = :id"), {"id": task_id}
-#         ).mappings().first()
-
-#     if row is None:
-#         raise HTTPException(status_code=404, detail="Task not found")
-
-#     return _task_from_row(row)
-
-# @app.delete("/tasks/{task_id}", status_code=204)
-# def delete_task(task_id: int) -> None:
-#     with get_connection() as connection:
-#         deleted = connection.execute(
-#             text("DELETE FROM tasks WHERE id = :id"), {"id": task_id}
-#         )
-#     if deleted.rowcount == 0:
-#         raise HTTPException(status_code=404, detail="Task not found")
-
-# @app.post("/tasks/{task_id}", response_model=Task, status_code=201)
-# def create_task222(task_id: int, payload: TaskCreate) -> Task:
-#     if payload.id != task_id:
-#         raise HTTPException(status_code=400, detail="Path task_id must match payload id")
-
-#     try:
-#         with get_connection() as connection:
-#             row = connection.execute(
-#                 text(
-#                     """
-#                     INSERT INTO tasks (id, name, task_group, employee, control, dependency, status, notifications, logs, comment)
-#                     VALUES (:id, :name, :task_group, :employee, :control, :dependency, :status, :notifications, :logs, :comment)
-#                     RETURNING *
-#                     """
-#                 ),
-#                 {
-#                     "id": payload.id,
-#                     "name": payload.name,
-#                     "task_group": payload.group,
-#                     "employee": payload.employee,
-#                     "control": payload.control,
-#                     "dependency": payload.dependency,
-#                     "status": payload.status,
-#                     "notifications": payload.notifications,
-#                     "logs": payload.logs,
-#                     "comment": payload.comment,
-#                 },
-#             ).mappings().one()
-#     except IntegrityError as exc:
-#         raise HTTPException(status_code=409, detail="Task with this id already exists or dependency is invalid") from exc
-
-#     return _task_from_row(row)
-
-
-# @app.put("/tasks/{task_id}", response_model=Task)
-# def replace_task(task_id: int, payload: TaskReplace) -> Task:
-#     with get_connection() as connection:
-#         row = connection.execute(
-#             text(
-#                 """
-#                 UPDATE tasks
-#                 SET name = :name,
-#                     task_group = :task_group,
-#                     employee = :employee,
-#                     control = :control,
-#                     dependency = :dependency,
-#                     status = :status,
-#                     notifications = :notifications,
-#                     logs = :logs,
-#                     comment = :comment
-#                 WHERE id = :id
-#                 RETURNING *
-#                 """
-#             ),
-#             {
-#                 "id": task_id,
-#                 "name": payload.name,
-#                 "task_group": payload.group,
-#                 "employee": payload.employee,
-#                 "control": payload.control,
-#                 "dependency": payload.dependency,
-#                 "status": payload.status,
-#                 "notifications": payload.notifications,
-#                 "logs": payload.logs,
-#                 "comment": payload.comment,
-#             },
-#         ).mappings().first()
-
-#     if row is None:
-#         raise HTTPException(status_code=404, detail="Task not found")
-
-#     return _task_from_row(row)
-
-# def _task_from_row(row: dict) -> Task:
-#     return Task(
-#         id=int(row["id"]),
-#         name=row["name"],
-#         group=row["task_group"],
-#         employee=row["employee"],
-#         control=row["control"],
-#         dependency=row["dependency"],
-#         status=row["status"],
-#         notifications=row["notifications"],
-#         logs=row["logs"],
-#         comment=row["comment"],
-#     )
